Added documentation, and modified the layout of files
