+++ app/src/main/java/com/android/feedme/MainActivity.kt
@@ -12,8 +12,8 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.semantics.testTag
-import com.android.feedme.auth.LoginScreen
 import com.android.feedme.resources.C
+import com.android.feedme.ui.auth.LoginScreen
 import com.android.feedme.ui.camera.CameraScreen
 import com.android.feedme.ui.theme.feedmeAppTheme
